+++ analysis/src/nanoAnalysis.h
@@ -10,16 +10,16 @@
 
 #include <TString.h>
 
-#include "pileUpTool.h"
-#include "RoccoR.h"
-#include "lumiTool.h"
+#include "nano/analysis/external/pileUpTool.h"
+#include "nano/analysis/external/RoccoR.h"
+#include "nano/analysis/external/lumiTool.h"
 
-#include "MuonScaleFactorEvaluator.h"
-#include "ElecScaleFactorEvaluator.h"
+#include "nano/analysis/external/MuonScaleFactorEvaluator.h"
+#include "nano/analysis/external/ElecScaleFactorEvaluator.h"
 
-#include "BTagCalibrationStandalone.h"
-//#include "TopTriggerSF.h"
-//#include "TTbarModeDefs.h"
+#include "nano/analysis/external/BTagCalibrationStandalone.h"
+//#include "nano/analysis/external/TopTriggerSF.h"
+//#include "nano/analysis/external/TTbarModeDefs.h"
 
 #include "TMVA/Tools.h"
 #include "TMVA/Reader.h"
